cloud/requirements.txt
confluent_kafka[avro,json_schema]
python-dotenv
apache-flink>=1.20.0